Create fast and efficient model builder tool (#54)
### Description

This PR creates an initial version of a fast and efficient model builder
tool. It allows users to build an optimized and quantized ONNX model in
only a few minutes. The tool currently supports model architectures that
follow LLaMA-2, Mistral, and Phi-2.

### Motivation and Context

Currently, users have to first export large language models (LLMs) on
large machines with massive amounts of memory. Then, users have to run
the ORT transformer optimizer script to optimize the exported ONNX
model. Finally, users have to choose and run an ORT quantization script
to quantize the optimized ONNX model. For the state-of-the-art LLMs,
these steps are time-consuming and require machines with large amounts
of memory.

This tool combines all of these steps and greatly reduces both the time
taken and memory used. For example, generating a LLaMA-2 7B FP16
optimized ONNX model can take more than an hour. With this tool, it
takes only 3 minutes to generate the same model.