Separating test CMakeLists.txt from root CMakeLists.txt (#68)
This pull request primarily modifies the `CMakeLists.txt` file to
improve code organization and enable testing. The changes include the
addition of a new option to enable tests, the reorganization of test
related code into a separate `CMakeLists.txt` file in the `test`
directory, and the replacement of hardcoded paths with variables for
better maintainability.

Code organization and testing:

*
[`CMakeLists.txt`](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aR10):
Added a new option `ENABLE_TESTS` to enable or disable tests.
*
[`CMakeLists.txt`](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL35-R38):
Removed the hard-coded paths for the `ORT_HEADER_DIR` and `ORT_LIB_DIR`
and replaced them with variables. This makes the code more maintainable
and less prone to errors.
[[1]](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL35-R38)
[[2]](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL89-R96)
[[3]](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL157-R155)
*
[`CMakeLists.txt`](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL117-L135):
Removed the `unit_tests` target and related code from this file.
[[1]](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL117-L135)
[[2]](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL169)
[[3]](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL214-L235)
*
[`test/CMakeLists.txt`](diffhunk://#diff-33394812ba204689144fd2f80832db83853ba1cb32403edb4e15fe4893e675fdR1-R38):
Added a new `CMakeLists.txt` file in the `test` directory. This file now
contains all the test related code, making the main `CMakeLists.txt`
file cleaner and easier to read. This also makes it easier to manage
test related code separately from the main build code.
*
[`CMakeLists.txt`](diffhunk://#diff-1e7de1ae2d059d21e1dd75d5812d5a34b0222cef273b7c3a2af62eb747f9d20aL53-L57):
Removed the `test_srcs` from the `file(GLOB` command as the test sources
are now handled separately in the `test/CMakeLists.txt` file.