Separating python CMakeLists.txt from root CMakeLists.txt (#69)
This pull request primarily involves changes in the `CMakeLists.txt`
file and the addition of a new `CMakeLists.txt` in the `src/python`
directory. The changes are mainly about refactoring the build process
for the Python module and its dependencies, including the Tokenizer. The
changes also include the relocation of the Python wheel building process
to the new `CMakeLists.txt` file in the `src/python` directory.

Key changes:

Refactoring in `CMakeLists.txt`:

* Removed the `PYTHON_ROOT` setting.
* Removed the condition to add the Tokenizer subdirectory and the
corresponding message in the `add_library(onnxruntime-genai-static
STATIC ${generator_srcs})` section.
* Replaced the condition to add the Tokenizer subdirectory and the
corresponding message to the `else()` section. Also, the message for
enabling tests has been updated with a more noticeable format.
* Removed the condition to set properties and link libraries for the
Python module in the `if(USE_CUDA AND CMAKE_CUDA_COMPILER)` section.
* Removed the condition to include directories, link libraries for the
Python module, and the message for linking the Tokenizer in the
`if(USE_TOKENIZER)` section. Also, the entire section for building the
Python wheel has been removed.

Addition in `src/python/CMakeLists.txt`:

* A new `CMakeLists.txt` file has been added in the `src/python`
directory. This file contains the build process for the Python module
and its dependencies, as well as the Python wheel building process.