Enhance model builder tool for building and running models (#107)
### Description

This PR enhances the model builder tool for building and running models
by adding the following.

- Adds a README file to show how to use the tool
- Creates a GenAI JSON config file that users can customize
- Saves the tokenization files from Hugging Face
- Adds additional options for customizing how the model is built
- Allows mini models to be created for testing purposes
- Rewrites `create_model` for integration with Olive

### Motivation and Context

- The README file showcases the tool's full usage.
- The JSON file has two groups, `model` and `search`. The `model`
section contains information about the model and the parameters in the
`search` section can be modified as desired.
- The tokenization files are needed for the tokenizers in the ORT GenAI
library.
- The full list of additional options can be viewed in the usage section
of the README.
- By reducing the number of hidden layers, mini models can be created
with the model builder tool for testing purposes (e.g. unit testing,
CIs).
- [This PR](https://github.com/microsoft/Olive/pull/940) is integrating
the model builder tool into Olive and wants to import the `create_model`
method directly instead of running the script.