Create CAPI publishing pipleine for Windows (#154)
This pull request includes significant changes to the `.pipelines`
directory, with the goal of improving the flexibility and
maintainability of the pipeline configuration. The most important
changes include the addition of dynamic parameters to various pipeline
stages, the consolidation of packaging stages, and the renaming of jobs
for clarity.

Changes to pipeline configuration:

*
[`.pipelines/capi-publishing.yml`](diffhunk://#diff-a3d5aa9136afcaee892e0e3e8f07a4c82b0071f03da920c065ba4d981778beaeL55-L76):
The pipeline configuration has been simplified by consolidating the
different packaging stages into a single template,
`capi-packaging-stage.yml`. This change also introduces dynamic
parameters for enabling different architectures and platforms.
*
[`.pipelines/nuget-publishing.yml`](diffhunk://#diff-e6ec5ea53c5d408941757974cb3e698481ef8f5bf36870475df104ef6e7b3277L63-R63):
The `ort_version` parameter has been made dynamic, allowing it to be set
externally.

Addition of dynamic parameters:

*
[`.pipelines/stages/capi-packaging-stage.yml`](diffhunk://#diff-55023ddc7643476268666d958caad9b8a76ea109e8dde85c8f37b8eb58dd841dR1-R42):
This new file defines a stage for packaging the C API. It introduces
dynamic parameters for enabling different architectures and platforms,
as well as for setting the `ort_version` and `cuda_version`.
*
[`.pipelines/stages/nuget-packaging-stage.yml`](diffhunk://#diff-1e491a28b1b87f333551d4804a6ab37ab1edcfdee6968e4b7d92f60f4f285e14R22-R42):
This file has been updated to use dynamic parameters for enabling
different architectures and platforms.

Job renaming and parameter changes:

*
[`.pipelines/stages/jobs/capi-win-packaging-job.yml`](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9L2-R34):
The job has been renamed to `Windows_CAPI_Packaging_${{ parameters.arch
}}_${{ parameters.ep }}` for clarity. It also introduces dynamic
parameters for the architecture (`arch`) and execution provider (`ep`).
[[1]](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9L2-R34)
[[2]](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9L57-R72)
*
[`.pipelines/stages/jobs/nuget-win-packaging-job.yml`](diffhunk://#diff-97390677658d875cbd41773aa78037661f27d91cd3d878db42a9073d3764db51L4-R27):
The job has been renamed to `Windows_Nuget_Packaging_${{ parameters.arch
}}_${{ parameters.ep }}` for clarity. It also introduces dynamic
parameters for the architecture (`arch`) and execution provider (`ep`).
[[1]](diffhunk://#diff-97390677658d875cbd41773aa78037661f27d91cd3d878db42a9073d3764db51L4-R27)
[[2]](diffhunk://#diff-97390677658d875cbd41773aa78037661f27d91cd3d878db42a9073d3764db51R37)