ThirdPartyNotices (#171)
This pull request mainly focuses on adding the `ThirdPartyNotices.txt`
file to different parts of the codebase. Specifically, the changes
involve including this file in the build artifacts, NuGet packages, and
Python wheel build.

Here are the most significant changes:

*
[`.pipelines/stages/jobs/capi-win-packaging-job.yml`](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9R69):
Added the `ThirdPartyNotices.txt` file to the list of files included in
the build artifacts.
* `nuget/Microsoft.ML.OnnxRuntimeGenAI.Managed.nuspec` and
`nuget/Microsoft.ML.OnnxRuntimeGenAI.nuspec`: Included the
`ThirdPartyNotices.txt` file in the NuGet packages.
[[1]](diffhunk://#diff-01d2fd48e40c38ac82de2127f5fb24b552ed6d9716b495f3dd1cfbd0f900a87aR22)
[[2]](diffhunk://#diff-3c5ea4681b1c3e5e5af7c7fc811052a2d530356e8fbfe244c5429077d36d5778R32)
*
[`src/python/CMakeLists.txt`](diffhunk://#diff-43c9042647d14a73e4061f77bba19ee368fd059a7e0c6086063f7646e2b2aabbR54):
Added the `ThirdPartyNotices.txt` file to the Python wheel build.