Remove clang from CI pipeline. (#182)
This pull request includes changes to the
`.github/workflows/linux-cpu-x64-build.yml` file. The changes simplify
the build process by removing the matrix strategy for different
compilers and standardizing on the gcc compiler. This has resulted in
changes to the build and test commands, and the removal of an artifact
upload step that was specific to the gcc build.

Changes to build process:

* Removed the matrix strategy for different compilers, standardizing on
the gcc compiler. This change affects the `env:` section of the `jobs:`
block.
* Updated the build and test commands to remove references to the matrix
strategy and use the gcc compiler directly. This change is reflected in
several places in the `jobs:` block.
[[1]](diffhunk://#diff-e758ed36b14737d7c89353329cb735a6d081f5fa288b809d49e47bfc0ade96c3L40-R42)
[[2]](diffhunk://#diff-e758ed36b14737d7c89353329cb735a6d081f5fa288b809d49e47bfc0ade96c3L63-R65)
* Removed the artifact upload step that was specific to the gcc build.
This change is also in the `jobs:` block.
* Clang will be used for nightly build