Enable C API packaging for linux (#178)
This pull request includes changes to several pipeline and job files,
primarily to remove support for Windows ARM and Linux ARM builds, and to
update the default settings for Linux CPU and GPU builds. Additionally,
there are changes to the parameters for several jobs and the addition of
new parameters.

Changes to pipeline files:

*
[`.pipelines/capi-publishing.yml`](diffhunk://#diff-a3d5aa9136afcaee892e0e3e8f07a4c82b0071f03da920c065ba4d981778beaeL12-R20):
Removed parameters for Windows ARM and Linux ARM builds, and changed the
default settings for Linux CPU and GPU builds.
[[1]](diffhunk://#diff-a3d5aa9136afcaee892e0e3e8f07a4c82b0071f03da920c065ba4d981778beaeL12-R20)
[[2]](diffhunk://#diff-a3d5aa9136afcaee892e0e3e8f07a4c82b0071f03da920c065ba4d981778beaeL59-L62)
*
[`.pipelines/nuget-publishing.yml`](diffhunk://#diff-e6ec5ea53c5d408941757974cb3e698481ef8f5bf36870475df104ef6e7b3277L12-L16):
Removed parameters for Windows ARM and Linux ARM builds.
[[1]](diffhunk://#diff-e6ec5ea53c5d408941757974cb3e698481ef8f5bf36870475df104ef6e7b3277L12-L16)
[[2]](diffhunk://#diff-e6ec5ea53c5d408941757974cb3e698481ef8f5bf36870475df104ef6e7b3277L27-L30)
[[3]](diffhunk://#diff-e6ec5ea53c5d408941757974cb3e698481ef8f5bf36870475df104ef6e7b3277L57-L62)
*
[`.pipelines/pypl-publishing.yml`](diffhunk://#diff-e7c846e17ab5319529d839ddce86308159f05bc27ed172b501c1b996d00d838cL12-L16):
Removed parameters for Windows ARM and Linux ARM builds, and added a new
parameter for CUDA version.
[[1]](diffhunk://#diff-e7c846e17ab5319529d839ddce86308159f05bc27ed172b501c1b996d00d838cL12-L16)
[[2]](diffhunk://#diff-e7c846e17ab5319529d839ddce86308159f05bc27ed172b501c1b996d00d838cL27-R34)
[[3]](diffhunk://#diff-e7c846e17ab5319529d839ddce86308159f05bc27ed172b501c1b996d00d838cL50-R52)

Changes to job files:

*
[`.pipelines/stages/capi-packaging-stage.yml`](diffhunk://#diff-55023ddc7643476268666d958caad9b8a76ea109e8dde85c8f37b8eb58dd841dL6-L42):
Removed parameters for Windows ARM and Linux ARM builds, and updated the
jobs for Windows and Linux packaging.
*
[`.pipelines/stages/jobs/capi-linux-packaging-job.yml`](diffhunk://#diff-dd3d7a97362e7af29378680a0a69aacc368e85fa3d6f4fc32dfb48e2514b289fR1-R94):
Added new parameters for architecture, endpoint, ONNX Runtime version,
and CUDA version, and updated the job steps.
*
[`.pipelines/stages/jobs/capi-win-packaging-job.yml`](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9R10-R36):
Added a new default parameter for CUDA version, updated the artifact
name and build directory variables, and updated the job steps.
[[1]](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9R10-R36)
[[2]](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9L40-R64)
[[3]](diffhunk://#diff-38b1b9a2465d43797e09697e900eba44936087f430fc3115b49209f5621b31b9L70-R90)
*
[`.pipelines/stages/jobs/nuget-win-packaging-job.yml`](diffhunk://#diff-97390677658d875cbd41773aa78037661f27d91cd3d878db42a9073d3764db51R10-R11):
Added a new default parameter for CUDA version, updated the build
directory variable, and updated the job steps.
[[1]](diffhunk://#diff-97390677658d875cbd41773aa78037661f27d91cd3d878db42a9073d3764db51R10-R11)
[[2]](diffhunk://#diff-97390677658d875cbd41773aa78037661f27d91cd3d878db42a9073d3764db51L19-R46)
*
[`.pipelines/stages/jobs/py-linux-cpu-packaging-job.yml`](diffhunk://#diff-da741590402f2646d68481d93e55f810d2140147c06942438ca8ce480273dbceL1-L167):
Removed the entire job file.