Add a random_seed search option (#250)
It goes into the search options as `"random_seed":1234` (for example).
The default value is -1, which means to use a random seed.

There is an issue with CUDA where the output can eventually differ even
if the same seed is used. I did some investigation and the random
numbers match, but the tokens chosen will differ eventually in longer
outputs.

With CPU the output always matches.