Make inputs and outputs more flexible in model builder (#261)
### Description

This PR adds support for excluding the embedding layer and the language
modeling head when requested. It also refactors how the inputs and
outputs are structured within the model builder.

### Motivation and Context

There are internal scenarios where excluding the embedding layer or
excluding the language modeling head from the final ONNX model is
useful.