Destroy the generator at the end of the loop in python examples (#404)
In python, when a variable overwrites an older variable with the same
line (as with `generator = og.Generator(model, params)` in the
`model-qa.py` loop), the new object is constructed before the old object
is destructured. For most objects the order doesn't really matter, but
when using graph capture, it does. The reason is that generators hold a
reference on the captured graph in order to allow it to be re-used in
the future without having to re-capture it, and 2 generators cannot hold
a reference on the same graph at the same time since it would result in
race conditions.

To work around this issue, we can delete the generator at the end of the
loop to make sure the captured graph is freed before the next iteration.
Note that this is only a problem in python, as other languages (e.g. C#,
C++) will destroy the object as soon as it gets out of scope (even if
the C# garbage collector doesn't actually collect the memory at that
time, the object is destroyed and the reference on the captured graph is
released).

An alternative way to achieve the same thing would be to extract the
content of the inner loop into its own function. That way, the generator
would automatically be destroyed at the end of the function because
Python understand that it cannot possibly be reused.