Reduce zip package size and update zip package naming patten (#427)
This pull request includes changes to the
`.pipelines/stages/jobs/steps/utils/download-ort.yml` and
`cmake/package.cmake` files. The changes primarily involve the
modification of the `libonnxruntime.so` file handling and the
`CPACK_PACKAGE_FILE_NAME` setting.

The most important changes are:

*
[`.pipelines/stages/jobs/steps/utils/download-ort.yml`](diffhunk://#diff-b33e2529c1e7627822171d08975dbb8c52cdde1bcce3a5931a90ddf24c5019b1L50-R51):
The `libonnxruntime.so` file is now moved and renamed to
`libonnxruntime.so.$ort_stable_version` instead of being copied.
Additionally, a symbolic link is created from
`libonnxruntime.so.$ort_stable_version` to `libonnxruntime.so`. This
change improves the handling of the `libonnxruntime.so` file and its
versioning.

*
[`cmake/package.cmake`](diffhunk://#diff-5f5cbcf4c6721a0a3c7b347a598906a68658622f06352b6f8c07d45d1e11e0c4L16-R45):
The setting of `CPACK_PACKAGE_FILE_NAME` has been expanded to account
for different system processors and the use of CUDA or DML. This change
makes the package naming more specific and informative, allowing for
better identification of the appropriate package for each system
configuration.