Adding dependency libraries to zip package (#418)
This PR request primarily focuses on changes to adding ONNXRUNTIME
libraries to the zip packages. Depends on the Target EP, it could also
contain cuda_provider library or DML libraries.

Changes to `cmake/package.cmake`:

*
[`cmake/package.cmake`](diffhunk://#diff-5f5cbcf4c6721a0a3c7b347a598906a68658622f06352b6f8c07d45d1e11e0c4L7-R67):
Modified the `set_target_properties` to provide more specific
destination paths for the library, runtime, and public header. Also,
updated the `CPACK_PACKAGE_FILE_NAME` to be more specific based on the
system and processor type. This helps in creating more specific and
identifiable package names.
* 

Changes to `libonnxruntime.so` handling: To reduce linux package size. 
The major changes include replacing copy commands with symbolic links
for `libonnxruntime.so` and modifying the `set_target_properties` in
`cmake/package.cmake` to provide more specific destination paths and
package file names based on the system and processor type.

*
[`.github/workflows/linux-cpu-x64-build.yml`](diffhunk://#diff-e758ed36b14737d7c89353329cb735a6d081f5fa288b809d49e47bfc0ade96c3L67-R68):
Replaced the copy command with a symbolic link for `libonnxruntime.so`.
This creates a symbolic link to the original file instead of creating a
copy, saving disk space and ensuring any updates to the original file
are reflected in the linked file.
*
[`.github/workflows/linux-gpu-x64-build.yml`](diffhunk://#diff-e941545683ca21b146fc6f414d9fe9e311f5f5efc44fce47287e8b0d835434aeL81-R82):
Similar to the above change, replaced the copy command with a symbolic
link for `libonnxruntime.so`.
*
[`.pipelines/stages/jobs/steps/utils/download-ort.yml`](diffhunk://#diff-b33e2529c1e7627822171d08975dbb8c52cdde1bcce3a5931a90ddf24c5019b1L50-R50):
Again, replaced the copy command with a symbolic link for
`libonnxruntime.so`.

---------

Co-authored-by: Nat Kershaw <nakersha@microsoft.com>