Test fix - ensure random indices don't overlap (#445)
CreateRandomLogits might hit the same index twice in the random
sampling, this can result in num_large not being hit and resulting in
fewer large values in the logits. This will make our unit test fail
randomly.

For example, we have a TopK(10) test, and we randomly set 10 score
values to be 25 (the rest between 0-1). We test it by setting 10 random
values to be 25, and verify the TopK results are all '25'.

But, if a duplicate random index occurs, we'll only have 9 (or fewer)
score values as 25, and our Top 10 will include some lower values.