Cleanup Github CI (#434)
This pull request includes changes to the GitHub workflow files for the
Linux and Windows builds. The changes primarily focus on improving the
robustness of the scripts and updating the .NET version used in the
Windows builds.

Script Robustness Improvements:

*
[`.github/workflows/linux-cpu-x64-build.yml`](diffhunk://#diff-e758ed36b14737d7c89353329cb735a6d081f5fa288b809d49e47bfc0ade96c3R57):
Added `set -e -x` to the script in the `Extract OnnxRuntime library and
header files` job, which will cause the script to exit on the first
error and print each command before it's executed.
*
[`.github/workflows/linux-gpu-x64-build.yml`](diffhunk://#diff-e941545683ca21b146fc6f414d9fe9e311f5f5efc44fce47287e8b0d835434aeR71):
Similar to the CPU build, `set -e -x` was added to the script in the
`Extract OnnxRuntime library and header files` job.

.NET Version Update and Workflow Refactor:

*
[`.github/workflows/win-cpu-x64-build.yml`](diffhunk://#diff-14c154d7bf4d7c90f647475b0340d0840319223d91faa598c4756739555bd863L37-R49):
Updated the .NET version from `6.0.x` to `8.0.x` and refactored the
`Install jq and nuget` and `Download OnnxRuntime Nightly` jobs. The
`nuget install` command now uses the `$ORT_NIGHTLY_VERSION` variable
directly.
*
[`.github/workflows/win-cuda-x64-build.yml`](diffhunk://#diff-a210d203f64f630a798afd3cdb0f42547b71e46f668b055f563d2018d6379055L44-R56):
Similar to the CPU build, the .NET version was updated to `8.0.x` and
the same refactoring was applied to the `Install jq and curl` and
`Download OnnxRuntime Nightly` jobs.