Support extra inputs in graph captures (#464)
Since extra inputs can change between generators, we need a way to make
sure that a captured graph can be reused. For a graph to be reused, the
following conditions need to be met:

1. All extra inputs need to have the same names as the extra inputs that
were provided when first recording the graph
2. All shapes need to match
3. All data types need to match
4. The number of extra inputs needs to be the same

As long as those conditions are satisfied, we are allowed to reuse the
graph. Then, at model creation time, we need to copy the extra inputs to
the gpu-backed static buffers that were created when recording the
graph.