Add custom DML allocators and non-global env to work around memory issues (#467)
This change achieves 2 things:

1. Use non-global envs for DML only. It's also possible to enable it for
CUDA, but I wanted to avoid breaking things. I saw comments in the code
that indicated that CUDA somehow needs a global environment that
outlives the model, but it doesn't seem to be necessary for the DML EP.
Global envs tend to hold references on allocators, which makes it hard
to clean all memory when a model gets released. By having the lifetime
of the env be tied to the lifetime of the model, we avoid such problems.
2. Use a custom allocator for DML. `onnxrutime-genai-directml` is a very
specific scenario where we know in advance where all memory will be
coming from, so we don't need a fancy arena allocator to manage the
memory. In fact, having the memory as tight as possible allows models to
run on more hardware and reduces fragmentation.