Update base image and upgrade GCC version (#548)
1. Use an internal prebuilt base image instead of a public image, so
that we do not need to rebuild the manylinux part again and again.
2. The new CUDA 11 base image is based on almalinux 8 instead of ubi8,
so that we can get GCC 11. ubi8 only has GCC 8 and GCC 12, but GCC 12 is
compatible with CUDA 11. So before this change we use GCC 8 in CUDA 11
build. After this change we will use GCC 11 instead.
3. Drop the support for GCC 10 and below.

This PR provides another solution for #257 .