Add large language models to CI pipelines (#780)
### Description

This PR adds several large-language models to the CI pipelines.

### Motivation and Context

This allows us to run end-to-end tests for supported models to ensure
they work as new changes are made. Changes from [this
PR](https://github.com/microsoft/onnxruntime-genai/pull/774) are used
for some of the quantized PyTorch models to run correctly.