Add re-designed support for Phi-3 vision and Phi-3.5 vision (#882)
### Description
This PR supports the re-designed export of Phi-3 vision and Phi-3.5
vision. The new design natively supports multi-image and the `select`
logic inside the ONNX models.

### Motivation and Context
With the re-designed export, some of the logic inside ONNX Runtime GenAI
is no longer needed as it is now inside the ONNX model. This allows
other models to more easily re-use the vision and embedding components
within ONNX Runtime GenAI.