Fix destructor mismatch that caused random heap corruption (#972)
I don't know how this ever worked, but it's totally fair that I was the
one to suffer for it since I wrote the original line of code.

In something I was working on, genai would crash randomly with heap
corruption. I finally figured out it was this line that was the culprit.

I'll think about renaming things so nobody else gets confused like I
did.