Fix buffer overrun on Linux (#1010)
`cstr_buffer[length] = 0;` was writing past the end of the buffer as
length was already including the null.

Rewrote it so there's just one simpler codepath on all OS's. Plus memcpy
is well understood as it avoids the 'is null included' issue entirely.

Thanks to @aciddelgado & @baijumeswani for narrowing down the cause.