Fix past_present_share_buffer=false kv cache bug and asserts (#1110)
This PR fixes a bug where, when using continuous decoding, the output
would differ depending on whether past_present_share_buffer was on or
off. The bug was actually deeper than this, as the KV-Cache was missing
the last generated token when using continuous decoding, regardless of
share buffer.

It also fixes a few assert statements.