implement fix for set inputs (#1152)
### Fix SetInputs for Multi-Modal Models

#### Description
This PR addresses an issue where `SetInputs`, which is used with
multi-modal models like phi3v and whisper, was broken. The error
encountered was:

```
RuntimeError: Please use params.SetInputs for phi3v. AppendTokens is not supported for this model type.
```

#### Root Cause
The error was caused because `AppendTokens` is an API only meant to be
used by models that support it. However, in the case of multi modal
models, it was called internally during the construction of the
generator to process inputs, leading to the runtime error.

#### Fix
This PR avoids calling `AppendTokens` internally, thereby fixing the
issue with `SetInputs` on multi-modal models. It also fixes this issue:
https://github.com/microsoft/onnxruntime-genai/issues/1151.