Python array to OrtValue could fail due to memory ownership issues (#1225)
The previous method would reference the python memory, which can be
deleted and cause the OrtValue to now refer to deleted memory. This now
makes a copy of the memory into the OrtValue to prevent this from
happening.

I also added a check to ensure the memory is contiguous, as we do not
support non-contiguous memory yet. The user will see the error if they
don't, and they can easily fix this from the python side if they do hit
this.

Fix verified by Kunal (as he was hitting a crash caused by the previous
version)