Fix iOS packaging pipeline after static library removal (#1316)
After discussing with a few teammates who worked on the iOS packaging,
switching to a shared library is fine but the steps below were removed
as we no longer need to hide the symbols as it's no longer a static
library. So now it's just a simple step of copying the shared library
into the package.