Add value_stats logging option to show statistical information about … (#1352)
Adds a new logging option "value_stats" that is on by default. This will
show 'value_stats' after any array of values is also logged.

So when tensor inputs/output data is shown, the 'value_stats' will be
shown after them. Also for things like 'model_logits' as below:

```
  model_logits   Values[ 2.51172 4.83594 6.32422 1.30469 1.97949 ... -1.28418 -1.28418 -1.28711 -1.28516 -1.28223 ]
  value_stats    Min: -5.35156 Max: 8.35156 Mean: 0.933448 StdDev: 1.1297
```


![image](https://github.com/user-attachments/assets/c5e0ae84-8602-4001-9560-53fc10ae759e)

This will also detect non-finite values and say the index of the first
value similar to this image: (was renamed to just 'value_stats' but the
rest is accurate)


![image](https://github.com/user-attachments/assets/ebb06b0c-1402-4cd1-b2d4-2e1bbd67b1e0)


This also generalizes the Float16ToFloat32 code to be templatized on the
number of bits per field. This makes the implementation easier to read
(no more magic numbers) and also adds bfloat16 support trivially.